move logic to special class
